* ThreadSearch: commit forgotten format-fix and typo-correction
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8321 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/ThreadSearch/TextFileSearcher.cpp
@@ -149,7 +149,7 @@ TextFileSearcher::eFileSearcherReturn TextFileSearcher::FindInFile(const wxStrin
                 line.Trim(false);
                 line.Trim(true);
 
-                foundLines.Add(wxString::Format(wxT("%d"), i + 1));
+                foundLines.Add(wxString::Format(wxT("%lu"), static_cast<unsigned long>(i + 1)));
                 foundLines.Add(line);
             }
         }
